server/requirements.txt
fastapi
uvicorn[standard]
pyyaml
bcrypt
python-jose[cryptography]
